+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-...
+
+### Added
+
+- Template filter `keep_whitespace` to preserve whitespace in values from the database within the `clean_plaintext` filter block.
 
 ## [v1.0.2] - 2026-02-21